MAKE-1331: Clean up change point related work (#233)
